test(scoring): collapse dual-tree scoring specs into shared suites
The four scoring spec families (model, profiles, config, engine) existed
twice — once in packages/dns-checks/src/__tests__/scoring/ (run against the
SOURCE module) and once in test/ (run against the BUILT @blackveil/dns-checks
package). The dual run is intentional src↔dist/DTS drift coverage, but the
assertions + expected values were hand-mirrored across both trees and had
already drifted (the BUILT surface was missing several assertions the source
surface carried, and vice versa).

Extract each family's assertions into a single shared
scoring-<name>.suite.ts exporting defineScoring<Name>Suite(module); both
spec files become 9-line thin wrappers that inject their import surface
(source vs built) and call the suite. Logic + expected values now live once;
both surfaces still run, so drift is still caught — and parity is restored
(the union of both trees' previously-unique assertions now runs on each).

The .suite.ts files are not .spec.ts/.test.ts, so neither vitest run collects
them directly. They never ship: packages/dns-checks publishes only `dist`
(files allowlist; verified via npm pack --dry-run), so the vitest import in
src/__tests__ stays test-only.

CLAUDE.md step 19 updated to point maintainers at the .suite.ts, not the
wrappers (the old "both exist twice, update all four" guidance was now a trap).

Verified: package scoring dir 130 pass (6 files), root scoring specs 96 pass
(4 files) — parity restored; root + package typecheck clean; lint clean.
