+++ CLAUDE.md
@@ -217,7 +217,7 @@ Pattern-based, emits `fuzzing_suspected` to `ALERT_WEBHOOK_URL` from 15-min cron
 16. **New `ToolRuntimeOptions` field** for a binding: extend `BvMcpEnv` AND populate at all 3 construction sites in `src/index.ts` — else `ro.<field>` is undefined and the tool returns `{ unprovisioned: true }`
 17. **Count surfaces (single tripwire)** — the tool count is hardcoded in exactly ONE place: bump `EXPECTED_TOOL_COUNT` in `test/audits/tool-count-ssot.audit.test.ts`. Every other count **derives** from `TOOLS`/`TOOL_DEFS` (the specs assert against `TOOLS.length`, not literals) — do NOT re-introduce a magic `toHaveLength(N)`. Two membership tripwires remain: (a) if `scanIncluded`, add the tool to `EXPECTED_SCAN_DOMAIN_TOOLS` in `tool-schemas.spec.ts` (exact-set guard on the scoring-critical scan set); (b) a non-CheckResult (custom-shape) tool goes in `NON_CHECK_RESULT_TOOLS` in `src/schemas/tool-definitions.ts` (the exported SSOT set that drives `outputSchema` population — no separate count to bump)
 18. **Audits that fail in CI** — `tool-count-ssot` (the count tripwire, item 17); `server-json-tool-count` (server.json "N MCP tools" vs `TOOLS.length`); `readme-tool-surface` (README + `docs/github-settings.md` + `extensions/vscode/{README.md,package.json}` prose must match the derived `TOOLS.length`, incl. the `check_*` count — update the prose strings); `npm run generate:wasm-permissions` (generated Rust perms); `test/chaos/varied-domain-all-tools.chaos.test.ts` (case list = registry)
-19. **Scored only** — bump `scoring-model.spec.ts` (`CATEGORY_TIERS` length + per-tier counts) and update the score snapshots in `scoring-profiles.spec.ts`; **both exist twice** (`packages/dns-checks/src/__tests__/scoring/` AND `test/`). If scanned, also bump the scan-category count in `test/scan-domain.spec.ts`
+19. **Scored only** — bump the `CATEGORY_TIERS` length + per-tier counts and the score snapshots. All four scoring spec families (`model`, `profiles`, `config`, `engine`) now keep their assertions + expected values **once** in a shared `packages/dns-checks/src/__tests__/scoring/scoring-<name>.suite.ts`; the eight `scoring-<name>.spec.ts` (one per name in `packages/dns-checks/src/__tests__/scoring/` + one in `test/`) are 9-line thin wrappers that inject the SOURCE vs the BUILT `@blackveil/dns-checks/scoring` module — **edit the `.suite.ts`, not the wrappers**. If scanned, also bump the scan-category count in `test/scan-domain.spec.ts`
 
 ## Testing
